Stub in UserClient and extract common client fns to Client
