+++ pysyncgateway/__init__.py
@@ -4,10 +4,12 @@ from .admin_client import AdminClient
 from .database import Database
 from .document import Document
 from .user import User
+from .user_client import UserClient
 
 __all_ = [
     AdminClient,
     Database,
     Document,
     User,
+    UserClient,
 ]
