+++ pysyncgateway/client.py
@@ -1,0 +1,46 @@
+from __future__ import absolute_import, print_function, unicode_literals
+
+from requests import delete, get, put
+
+from .helpers import ComparableMixin, sg_method
+
+
+class Client(object, ComparableMixin):
+    """
+    Abstract parent class for `AdminClient` and `UserClient`.
+
+    Attributes:
+        url (str): Sync Gateway REST API URL.
+    """
+    CREATED = 1
+    UPDATED = 2
+    CONFLICT = 3
+
+    def __init__(self, url):
+        """
+        Args:
+            url (str): Sync Gateway REST API URL (excluding database but
+                including trailing slash).
+        """
+        self.url = url
+
+    def get_server(self):
+        """
+        Returns:
+            dict: Meta-information about the server.
+        """
+        return self.get(self.url).json()
+
+    # --- REST Verbs ---
+
+    @sg_method
+    def get(self, url, **kwargs):
+        return get(url, **kwargs)
+
+    @sg_method
+    def put(self, url, data):
+        return put(url, json=data)
+
+    @sg_method
+    def delete(self, url, **kwargs):
+        return delete(url, **kwargs)
