+++ pysyncgateway/user_client.py
@@ -1,0 +1,7 @@
+from __future__ import absolute_import, print_function, unicode_literals
+
+from .client import Client
+
+
+class UserClient(Client):
+    pass
