+++ tests/user_client/test_init.py
@@ -1,0 +1,9 @@
+from __future__ import absolute_import, print_function, unicode_literals
+
+from pysyncgateway import UserClient
+
+
+def test(syncgateway_public_url):
+    result = UserClient(syncgateway_public_url)
+
+    assert result.url == syncgateway_public_url
